+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Added
 - Util to regenerate OpenAPI test snapshots
-- Calling `cds compile -2 openapi` now emits `cds.compile.to.openapi` with an object parameter, containing the source model, the compilation options, and the compilation result.
+- Calling `cds compile -2 openapi` now emits events: `compile.to.openapi` with an object parameter, containing the source model and the compilation options before the compilation takes place. `after:compile.to.openapi` carries the same parameter with an additional `result` property, containing the output of the compilation process.
 
 ### Changed
 - Operations: `@Core.Description` maps to `summary`, `@Core.LongDescription` maps to `description`